修复文章分类删除问题：删除某分类时没有考虑其子分类会导致其子分类成为“悬挂”分类

+++ blog-core/src/main/java/com/zyd/blog/business/service/impl/BizTypeServiceImpl.java
@@ -95,6 +95,11 @@ public class BizTypeServiceImpl implements BizTypeService {
         if (!CollectionUtils.isEmpty(articles)) {
             throw new ZhydException("当前分类下存在文章信息，禁止删除！");
         }
+
+        boolean b = listAll().stream().anyMatch(a -> primaryKey.equals(a.getParentId()));
+        if (b) {
+            throw new ZhydException("当前分类下存在子分类，禁止删除");
+        }
         return bizTypeMapper.deleteByPrimaryKey(primaryKey) > 0;
     }
